requirements.txt
altgraph==0.17.5
charset-normalizer==3.4.4
colorama==0.4.6
MouseInfo==0.1.3
mss==9.0.1
opencv-python==4.9.0.80
packaging==25.0
pefile==2024.8.26
pillow==12.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
pywin32-ctypes==0.2.3
img2pdf==0.4.4
reportlab==4.4.7
setuptools==80.9.0
typing_extensions==4.15.0
ttkthemes==3.2.2
keyboard==0.13.5
numpy==1.26.4